+++ frontend/src/components/FeedingList/feeding-list.css
@@ -1,17 +1,17 @@
-.teacher-list {
-    width: 900px;
-    box-shadow: 0px 2px 3px rgb(199, 199, 199);
-    padding: 10px;
-    margin: 10px 0px;
-}
+/*.teacher-list {*/
+/*    width: 900px;*/
+/*    box-shadow: 0px 2px 3px rgb(199, 199, 199);*/
+/*    padding: 10px;*/
+/*    margin: 10px 0px;*/
+/*}*/
 
-.add-search-container {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 20px;
-}
+/*.add-search-container {*/
+/*    display: flex;*/
+/*    justify-content: space-between;*/
+/*    margin-bottom: 20px;*/
+/*}*/
 
-.del-update-container {
-    display: flex;
-    justify-content: space-between;
-}
+/*.del-update-container {*/
+/*    display: flex;*/
+/*    justify-content: space-between;*/
+/*}*/
