+++ frontend/src/components/PoemList/index.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import { Table, Button, Avatar, Popconfirm, Modal, Select } from "antd";
 import axios from "axios";
-import "./poem-list.css";
 import PoemForm from '../PoemForm';
 import { Link } from 'react-router-dom';
 import { useParams } from "react-router-dom";
