+++ frontend/src/index.js
@@ -1,23 +1,22 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+import { AuthProvider } from "./hooks/useAuth";
 
-import { BrowserRouter,createBrowserRouter } from 'react-router-dom';
+import App from "./App";
 
-// const router = createBrowserRouter([
-//     {
-//         path: "/",
-//         element: <Root />,
-//         errorElement: <ErrorPage />,
-//     },
-// ]);
+const rootElement = document.getElementById("root");
+const root = createRoot(rootElement);
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+
+
 root.render(
-    <React.StrictMode>
+    <StrictMode>
         <BrowserRouter>
-            <App />
+            <AuthProvider>
+                    <App />
+            </AuthProvider>
         </BrowserRouter>
-    </React.StrictMode>
+    </StrictMode>
 );
-
