--- frontend/src/page/Feeding/main.css
@@ -1,6 +1,0 @@
-.main-page-wrapper {
-    display: flex;
-    justify-content: center;
-    height: 100vh;
-    background-color: rgb(216, 216, 231);
-}
