+++ frontend/src/page/Home/index.js
@@ -10,18 +10,7 @@ class Home extends Component {
 
         return (
             <div>
-                <nav>
-                    <Link to="/">home</Link>
-                </nav>
-                <nav>
-                    <Link to="/Feeding">Feeding</Link>
-                </nav>
-                <nav>
-                    <Link to="/login">login</Link>
-                </nav>
-                <nav>
-                    <Link to="/register">register</Link>
-                </nav>
+              I am home
 
             </div>
         );
