+++ src/main/java/com/ssw555/babycare/Entity/User.java
@@ -18,6 +18,7 @@ public class User {
     private Integer id;
 
     private String username;
+    private Integer role;
     private String password;
